src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
alabaster=0.7.12=py37_0
appdirs=1.4.4=pyh9f0ad1d_0
argh=0.26.2=py37_0
arrow=0.13.1=py37_0
astroid=2.5=py37h06a4308_1
async_generator=1.10=py37h28b3542_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.6=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h88c068d_5
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.4=py37hc58025e_1
chardet=4.0.0=py37h89c1867_0
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
common_cmplr_lib_rt=2021.1.1=intel_189
common_cmplr_lic_rt=2021.1.1=intel_189
cookiecutter=1.7.2=pyhd3eb1b0_0
cryptography=3.3.1=py37h7f0c10b_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
cytoolz=0.11.0=py37h4abf009_1
dask-core=2020.12.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=py_0
docutils=0.16=py37_1
dpcpp_cpp_rt=2021.1.1=intel_189
entrypoints=0.3=py37_0
enum34=1.1.10=py37hc8dfbb8_2
expat=2.3.0=h2531618_2
ffmpeg=4.0=hcdf2ecd_0
flake8=3.9.0=pyhd3eb1b0_0
flow-vis=0.1=pypi_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.2.1=h9c3ff4c_2
freetype=2.10.4=h9e62b58_0
future=0.18.2=py37h89c1867_3
glib=2.68.1=h36276a3_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icc_rt=2021.1.1=intel_189
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
imagecodecs-lite=2019.12.3=py37h161383b_3
imageio=2.9.0=py_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
inflection=0.5.1=py37h06a4308_0
intel-openmp=2021.1.1=intel_189
intelpython=2021.1.1=1
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.8.0=pyhd3eb1b0_0
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=py37h06a4308_1
jeepney=0.6.0=pyhd3eb1b0_0
jinja2=2.11.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pyhd3eb1b0_2
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
keyring=22.3.0=py37h06a4308_0
kiwisolver=1.3.1=py37hc928c03_0
lazy-object-proxy=1.6.0=py37h27cfd23_0
lcms2=2.11=h396b838_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=3_h92ddd45_netlib
libffi=3.3=h07ac4c1_13
libgcc-ng=9.3.0=hdf63c60_101
libgfortran=3.0.0=1
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libglu=9.0.0=he1b5a44_1001
liblapack=3.9.0=3_h92ddd45_netlib
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h17b3f18_7
libsodium=1.0.16=3
libspatialindex=1.9.3=h2531618_0
libstdcxx-ng=9.3.0=hdf63c60_101
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=h58efa85_4
lz4-c=1.9.2=h7708b8d_3
markupsafe=1.1.1=py37h14c3975_1
matlabengineforpython=R2020a=pypi_0
matplotlib-base=3.3.2=py37h817c723_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h14c3975_1001
mkl=2021.1.1=intel_52
mkl-service=2.3.0=py37hfc5565f_6
mkl_fft=1.2.0=py37hb9c2cde_4
mkl_random=1.2.0=py37h90a4e38_4
mkl_umath=0.1.0=py37hb46b53c_0
mypy_extensions=0.4.3=py37_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h02626c5_0
numpy-base=1.19.2=py37h591eb60_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=py37_0
opencl_rt=2021.1.1=intel_189
opencv=3.4.2=py37h6fd60c2_1
openexr=1.3.2=pypi_0
openssl=1.1.1k=h7f98852_0
packaging=20.8=pyhd3deb0d_0
pandas=1.1.5=py37hdc94413_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.7.0=py_0
pathlib=1.0.1=py37hc8dfbb8_3
pathspec=0.7.0=py_0
patsy=0.5.1=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37he98fc37_0
pip=20.2.3=py37h56aae7b_1
pixman=0.40.0=h36c2ea0_0
pluggy=0.13.1=py37h06a4308_0
pooch=1.3.0=pyhd8ed1ab_0
poyo=0.5.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=py37h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=pyh9f0ad1d_2
pydocstyle=6.0.0=pyhd3eb1b0_0
pyflakes=2.2.0=pyhd3eb1b0_0
pygments=2.8.1=pyhd3eb1b0_0
pylint=2.7.4=py37h06a4308_1
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.2=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37he5f6b98_2
python=3.7.9=h7af079d_1
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytube3=9.6.4=pypi_0
pytz=2020.4=pyhd8ed1ab_0
pywavelets=1.1.1=py37h902c9e0_3
pyxdg=0.27=pyhd3eb1b0_0
pyyaml=5.3.1=py37hb5d75c8_1
pyzmq=18.1.1=py37he6710b0_0
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
regex=2021.4.4=py37h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
rope=0.18.0=py_0
rtree=0.9.7=py37h06a4308_1
scikit-image=0.18.0=py37hdc94413_0
scikit-learn=0.23.2=py37hddcf8d6_3
scikit-video=1.1.11=pyh24bf2e0_0
scipy=1.5.3=py37h14a347d_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
secretstorage=3.3.1=py37h06a4308_0
setuptools=50.3.2=py37h4200cf5_0
simplejson=3.17.2=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h65307dc_1
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=3.5.4=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.4=pyhd3eb1b0_0
spyder=5.0.0=py37h06a4308_1
spyder-kernels=2.0.1=py37h06a4308_0
sqlite=3.33.0=h88c068d_1
statsmodels=0.12.2=py37h902c9e0_0
tbb=2021.1.1=intel_119
tbb4py=2021.1.1=py37_intel_119
tcl=8.6.9=h14c3975_27
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=py_0
textdistance=4.2.1=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
three-merge=0.1.1=pyhd3eb1b0_0
tifffile=2019.7.26.2=py37_0
tinycss=0.4=pyhd3eb1b0_1002
tk=8.6.9=29
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.1=py_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu102
tornado=6.1=py37h4abf009_0
tqdm=4.54.1=pyhd8ed1ab_1
traitlets=5.0.5=pyhd3eb1b0_0
typed-ast=1.4.2=py37h27cfd23_1
typing_extensions=3.7.4.3=py_0
ujson=4.0.2=py37h2531618_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.2=pyhd8ed1ab_0
watchdog=1.0.2=py37h06a4308_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.35.1=py37h4a4c509_1
whichcraft=0.6.1=pyhd3eb1b0_0
wrapt=1.12.1=py37h7b6447c_1
wurlitzer=2.1.0=py37h06a4308_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.1=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=hcc43529_2
yaml=0.2.5=h516909a_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.1=he6710b0_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11.1=hb8a9d29_3
zstd=1.4.5=hdb51d2f_0
